# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Societal Innovation Portal API
  version: 1.0.0
  description: >
    Canonical machine-readable contract for the SIH 2026 Societal Innovation Portal.
    Human-facing notes, examples and socket events live in docs/API_CONTRACT.md.
    Request/response shapes are generated from packages/shared-types (Zod) --
    if the two disagree, packages/shared-types wins and this file is the bug.
servers:
  - url: http://localhost:4000
    description: Local development
  - url: https://report-maro-c8m1.onrender.com
    description: Deployed demo
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/projects:
    get:
      tags: [Projects]
      summary: Fundable projects (industry portal)
      security: []
      parameters:
        - { in: query, name: status, schema: { $ref: '#/components/schemas/ProjectStatus' } }
        - { in: query, name: category, schema: { $ref: '#/components/schemas/ProblemCategory' } }
        - { in: query, name: universityId, schema: { type: string } }
        - { in: query, name: fundedByMe, schema: { type: boolean } }
        - { in: query, name: page, schema: { type: integer, default: 1 } }
        - { in: query, name: pageSize, schema: { type: integer, default: 20 } }
      responses:
        '200':
          description: Paginated project summaries
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { const: true }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/ProjectSummary' }
                  meta: { $ref: '#/components/schemas/ListMeta' }
  /api/v1/projects/{id}:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string } }
    get:
      tags: [Projects]
      summary: Project detail with funding ledger
      security: []
      responses:
        '200':
          description: Project
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { const: true }
                  data: { $ref: '#/components/schemas/Project' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/projects/{id}/status:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string } }
    patch:
      tags: [Projects]
      summary: Advance or cancel a project (UNIVERSITY owner or ADMIN)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string, enum: [ACTIVE, COMPLETED, CANCELLED] }
                note: { type: string, maxLength: 500 }
      responses:
        '200':
          description: Updated project
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { const: true }
                  data: { $ref: '#/components/schemas/Project' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/projects/{id}/fundings:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string } }
    get:
      tags: [Projects]
      summary: Funding ledger for a project
      security: []
      responses:
        '200':
          description: Fundings, newest first
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { const: true }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Funding' }
                  meta: { $ref: '#/components/schemas/ListMeta' }
    post:
      tags: [Projects]
      summary: Pledge funding (INDUSTRY)
      description: >
        When the ledger total reaches budgetInr the project flips to ACTIVE and
        the linked problem to IN_PROGRESS. Over-funding beyond the budget is rejected.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amountInr]
              properties:
                amountInr: { type: integer, minimum: 1 }
                note: { type: string, maxLength: 500 }
      responses:
        '201':
          description: Project after the pledge was applied
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { const: true }
                  data: { $ref: '#/components/schemas/Project' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    internalKey:
      type: apiKey
      in: header
      name: x-internal-key
  responses:
...
    Forbidden:
      description: Authenticated but the role is not allowed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    NotFound:
      description: No such resource
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    Conflict:
      description: Illegal state transition or duplicate
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
  schemas:
    ApiError:
      type: object
      required: [success, error, requestId]
      properties:
        success: { const: false }
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum: [VALIDATION_ERROR, UNAUTHORIZED, FORBIDDEN, NOT_FOUND, CONFLICT, RATE_LIMITED, PAYLOAD_TOO_LARGE, INTERNAL_ERROR, SERVICE_UNAVAILABLE]
            message: { type: string }
            details:
              type: array
              items:
                type: object
                properties:
                  path: { type: string }
                  message: { type: string }
        requestId: { type: string }
    ListMeta:
      type: object
      properties:
        pagination:
          type: object
          required: [page, pageSize, total, totalPages]
          properties:
            page: { type: integer }
            pageSize: { type: integer }
            total: { type: integer }
            totalPages: { type: integer }
...
    ProblemCategory:
      type: string
      enum: [WATER_SANITATION, ROADS_TRANSPORT, EDUCATION, HEALTHCARE, ELECTRICITY, WASTE_MANAGEMENT, AGRICULTURE, PUBLIC_SAFETY, OTHER]
    Priority: { type: string, enum: [LOW, MEDIUM, HIGH, CRITICAL] }
    ProblemStatus:
      type: string
      enum: [SUBMITTED, PROCESSING, TRIAGED, DUPLICATE, ASSIGNED, IN_PROGRESS, RESOLVED, REJECTED]
...
    ProjectStatus: { type: string, enum: [AWAITING_FUNDING, ACTIVE, COMPLETED, CANCELLED] }
    OrganizationSummary:
      type: object
      required: [id, name, type, district]
      properties:
        id: { type: string }
        name: { type: string }
        type: { type: string, enum: [UNIVERSITY, INDUSTRY, GOVERNMENT] }
        district: { type: [string, 'null'] }
...
    Location:
      type: object
      required: [latitude, longitude, district]
      properties:
        latitude: { type: number, minimum: 21.5, maximum: 25.6 }
        longitude: { type: number, minimum: 83.2, maximum: 88.0 }
        district: { type: string }
        address: { type: string }
...
    ProblemSummary:
      type: object
      required: [id, title, status, location, reporter, category, priority, duplicateCount, assignmentId, createdAt, updatedAt]
      properties:
        id: { type: string }
        title: { type: string }
        status: { $ref: '#/components/schemas/ProblemStatus' }
        location: { $ref: '#/components/schemas/Location' }
        reporter:
          type: object
          properties:
            id: { type: string }
            name: { type: string }
        category:
          oneOf:
            - $ref: '#/components/schemas/ProblemCategory'
            - type: 'null'
        priority:
          oneOf:
            - $ref: '#/components/schemas/Priority'
            - type: 'null'
        duplicateCount: { type: integer }
        assignmentId: { type: [string, 'null'] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    Funding:
      type: object
      required: [id, projectId, industry, amountInr, note, fundedBy, createdAt]
      properties:
        id: { type: string }
        projectId: { type: string }
        industry: { $ref: '#/components/schemas/OrganizationSummary' }
        amountInr: { type: integer }
        note: { type: [string, 'null'] }
        fundedBy:
          type: object
          properties:
            id: { type: string }
            name: { type: string }
        createdAt: { type: string, format: date-time }
    ProjectSummary:
      type: object
      required: [id, proposalId, problem, university, title, category, status, budgetInr, fundedInr, fundedPercent, timelineWeeks, startedAt, completedAt, createdAt, updatedAt]
      properties:
        id: { type: string }
        proposalId: { type: string }
        problem: { $ref: '#/components/schemas/ProblemSummary' }
        university: { $ref: '#/components/schemas/OrganizationSummary' }
        title: { type: string }
        category:
          oneOf:
            - $ref: '#/components/schemas/ProblemCategory'
            - type: 'null'
        status: { $ref: '#/components/schemas/ProjectStatus' }
        budgetInr: { type: integer }
        fundedInr: { type: integer }
        fundedPercent: { type: number }
        timelineWeeks: { type: integer }
        startedAt: { type: [string, 'null'], format: date-time }
        completedAt: { type: [string, 'null'], format: date-time }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    Project:
      allOf:
        - $ref: '#/components/schemas/ProjectSummary'
        - type: object
          required: [summary, fundings]
          properties:
            summary: { type: string }
            fundings: { type: array, items: { $ref: '#/components/schemas/Funding' } }
...